==> 16_Feedback HUJI annotated.Rmd <==
---
title: "Feedback on R Workshop in English"
author: "Avi Kluger"
date: "February 14, 2019"
output:
  html_document: default
  word_document: default
  pdf_document: default
---

```{r echo=FALSE}

source("C://Users//owner//Dropbox//Rfunctions//Qualtrix//read_Qualtrics.R")
Sys.setenv("api_key"    = "UT0Dq7E6XR6heyNYWghyBigOA2Q4lwGELZieuyAz")
Sys.setenv("dataCenter" = "az1.qualtrics.com")

##### Read IL responses in English ####
df <- read_Qualtrics("SV_3eHBVUyJCMRX3hP")

df <- df[, grep("Satisfaction_1", names(df)):length(df)]

```

## R workshop feedback survey

At the end of the workshop, participants were asked to rate it and comment on it.  There were `r nrow(na.omit(df))` valid responses to the following questions: 

1. Overall, I am satisfied with the content of the R workshop [0 = Strongly disagree; 10 = Strongly Agree]
2. Overall, I am satisfied with the lecturer [0 = Strongly disagree; 10 = Strongly Agree]
3. Overall, I am satisfied with the assistant (Sarit Pery) [0 = Strongly disagree; 10 = Strongly Agree]
4. Please write comments


# Results


```{r warning=FALSE, out.width = "50%" }

df <- as.data.frame(df)
temp_df <- data.frame("Course Satisfaction" = numeric(), 
                      "Lecturer Satisfaction" = numeric(),
                      "Assistant Satisfaction (Sarit Pery)" = numeric(),
                      "Comments" = character(),
                       check.names = FALSE)

names(df) <- names(temp_df)

# The code below is not very efficient because it repeats 3 times, which is 
# about the limit above which I write a function or a loop.  This code is needed
# to force the charts to include values that were not used by any participants.

Q  <- data.frame(table(df[, 1]))
Q1 <- as.data.frame(cbind(0:10, !(0:10 %in% Q[, 1])))
names(Q) <- names(Q1)
Q1 <- merge(Q1, Q, by = "V1", all = TRUE)
Q1 <- Q1[, -2]
names(Q1) <- c("x", "y")
Q1$y[is.na(Q1$y)] <- 0

# demonstrate what the code does for Q1
Q1


Q  <- data.frame(table(df[, 2]))
Q2 <- as.data.frame(cbind(0:10, !(0:10 %in% Q[, 1])))
names(Q) <- names(Q2)
Q2 <- merge(Q2, Q, by = "V1", all = TRUE)
Q2 <- Q2[, -2]
names(Q2) <- c("x", "y")
Q2$y[is.na(Q2$y)] <- 0

Q  <- data.frame(table(df[, 3]))
Q3 <- as.data.frame(cbind(0:10, !(0:10 %in% Q[, 1])))
names(Q) <- names(Q3)
Q3 <- merge(Q3, Q, by = "V1", all = TRUE)
Q3 <- Q3[, -2]
names(Q3) <- c("x", "y")
Q3$y[is.na(Q3$y)] <- 0


# produce barcharts
library(ggplot2)

ggplot(Q1,aes(Q1$x, Q1$y)) +  
  geom_bar(stat = "identity") +  
  theme_bw() +
  scale_x_continuous(breaks=0:10, labels=c("Strongly\nDisagree", 
                                           1:9, "Strongly\nAgree" )) +
  xlab ("Response option") +
  ylab("Frequency") +
  scale_y_continuous(breaks = c(seq(0, max(Q1$y), by =2))) +
  theme(plot.title = element_text(hjust = 0.5)) +
  ggtitle("Frequencies of the responses to
          1. Overall, I am satisfied with the content of the R workshop\n") +
  annotate("text", x = 2, y = 7, 
           label = paste("  mean = ", round(mean(df[, 1]), 2),
                         "\n  SD = ", round(sd(df[, 1]), 2)))

ggplot(Q2,aes(Q2$x, Q2$y)) +  
  geom_bar(stat = "identity") +  
  theme_bw() +
  scale_x_continuous(breaks=0:10, labels=c("Strongly\nDisagree", 
                                           1:9, "Strongly\nAgree" )) +
  xlab ("Response option") +
  ylab("Frequency") +
  scale_y_continuous(breaks = c(seq(0, max(Q2$y), by =2))) +
  theme(plot.title = element_text(hjust = 0.5)) +
  ggtitle("Frequencies of the responses to
  2. Overall, I am satisfied with the lecturer\n") +
  annotate("text", x = 2, y = 12, 
           label = paste("  mean = ", round(mean(df[, 2]), 2),
                         "\n  SD = ", round(sd(df[, 2]), 2)))

ggplot(Q2,aes(Q3$x, Q3$y)) +  
  geom_bar(stat = "identity") +  
  theme_bw() +
  scale_x_continuous(breaks=0:10, labels=c("Strongly\nDisagree", 
                                           1:9, "Strongly\nAgree" )) +
  xlab ("Response option") +
  ylab("Frequency") +
  scale_y_continuous(breaks = c(seq(0, max(Q3$y), by =2))) +
  theme(plot.title = element_text(hjust = 0.5)) +
  ggtitle("Frequencies of the responses to
  3. Overall, I am satisfied with the assistant (Sarit Pery)\n") +
  annotate("text", x = 2, y = 12, 
           label = paste("  mean = ", round(mean(df[, 3]), 2),
                         "\n  SD = ", round(sd(df[, 3]), 2)))
```




```{r}
ttest <- t.test(df$`Course Satisfaction`, 
       df$`Lecturer Satisfaction`, 
       paired=TRUE, 
       conf.level=0.95)
round(ttest$statistic, 2)

ttest1 <- t.test(df$`Lecturer Satisfaction`, 
                 df$`Assistant Satisfaction (Sarit Pery)`, 
       paired=TRUE, 
       conf.level=0.95)
round(ttest1$statistic, 2)

```

Participants were more satisfied with the lecturer than with the course, as indicated by paired-sample *t*-test, *t*~`r ttest$parameter`~ = `r round( ttest$statistic, 2)`, *p* = `r round(ttest$p.value, 3)`.  Participants were marginally more satisfied with the assistant than with the lecturer, as indicated by paired-sample *t*-test, *t*~`r ttest1$parameter`~ = `r round( ttest1$statistic, 2)`, *p* = `r round(ttest1$p.value, 3)`

### 3. Comments that were not empty 

```{r }
library(knitr)
kable(df[4], col.names = c("verbatim"))
```


## Conclusions
Participants were generally very satisfied with the R workshop.  They also reported that they were somewhat less satisfied with the content.  Specifically, some participants indicated that spending more time on basic R concepts would have served them better. In addition, most participants would consider taking an advanced workshop in the summer.
